src/rest/requirements.txt
Django==5.1
djangorestframework==3.15.2
pymongo==4.4.1
django-cors-headers == 4.4.0
